refactor: extract download planning from main into a dedicated plan module
- Move DownloadPlan, plan_download_tasks, files_to_download, and
  local_path_key into src/plan.rs. Planning (URL building, name
  sanitization, collision detection) is a distinct concern from CLI
  orchestration; main.rs now only coordinates the run flow.
- Register the plan module in lib.rs and update the imports in
  main.rs accordingly.
- Update AGENTS.md Architecture section to show the actual
  downloader/ module tree (stream, retry, verify, signal, mtime)
  and the new plan.rs.
