+++ src/main.rs
@@ -8,20 +8,19 @@
 use clap::Parser;
 use colored::*;
 use ia_get::archive_metadata::{
-    XmlFile, XmlFiles, XmlMetadata, encode_download_path, fetch_and_parse_xml, get_xml_url,
-    save_xml_metadata, xml_file_name_of,
+    XmlFiles, XmlMetadata, fetch_and_parse_xml, get_xml_url, save_xml_metadata, xml_file_name_of,
 };
 use ia_get::constants::USER_AGENT;
 use ia_get::cookie::cookie_header_value;
-use ia_get::downloader::{self, DownloadTask};
+use ia_get::downloader;
+use ia_get::plan::{files_to_download, plan_download_tasks};
 use ia_get::utils::{
     create_spinner, finish_spinner, format_size, last_glyph, print_downloaded_line,
-    print_file_banner, sanitize_filename, validate_archive_url,
+    print_file_banner, validate_archive_url,
 };
 use ia_get::{IaGetError, Result};
 use indicatif::ProgressBar;
 use reqwest::{Client, Url};
-use std::collections::HashMap;
 use std::path::Path;
 use std::time::{Duration, SystemTime};
 
@@ -94,136 +93,6 @@ fn save_and_announce_xml(
     print_downloaded_line(last_glyph(), content.len() as u64, None);
 
     Ok(())
-}
-
-/// Filters out the archive's self-referencing `_files.xml` entry, whose
-/// checksum, mtime and size are unreliable, leaving the files to download.
-fn files_to_download(files: Vec<XmlFile>, xml_file_name: &str) -> Vec<XmlFile> {
-    files
-        .into_iter()
-        .filter(|file| file.name != xml_file_name)
-        .collect()
-}
-
-/// Normalised key for local-path collision detection.
-///
-/// On case-insensitive filesystems (Windows, default macOS) "a.pdf" and
-/// "A.pdf" are the same path, so the key is lowercased there; on
-/// case-sensitive filesystems (Linux) distinct casing stays distinct.
-fn local_path_key(path: &str) -> String {
-    #[cfg(target_os = "linux")]
-    {
-        path.to_string()
-    }
-    #[cfg(not(target_os = "linux"))]
-    {
-        path.to_lowercase()
-    }
-}
-
-/// The download plan: the tasks that will actually run, how many file
-/// names were sanitized, and the warning lines (sanitized, collided and
-/// skipped entries) that the caller prints.
-struct DownloadPlan {
-    tasks: Vec<DownloadTask>,
-    sanitized_count: usize,
-    warnings: Vec<String>,
-}
-
-/// Converts the parsed metadata into download tasks: builds each file's
-/// absolute URL and its sanitized local path, collecting a warning line
-/// for every rename, collision or skip.
-///
-/// Entries whose name encodes to an empty URL path (an empty name, or
-/// slashes only) are skipped: joining `""` would resolve to the metadata
-/// URL itself. Likewise, an entry whose sanitized local path collides with
-/// an earlier entry's is skipped, so one file never overwrites another.
-/// Collisions are compared case-insensitively on case-insensitive
-/// filesystems (Windows, default macOS), where "a.pdf" and "A.pdf" are the
-/// same path.
-///
-/// A failed URL join aborts the run: silently keeping the base URL would
-/// download the metadata file under the file's name.
-fn plan_download_tasks(files: Vec<XmlFile>, base_url: &Url) -> Result<DownloadPlan> {
-    let mut sanitized_count = 0;
-    let mut tasks: Vec<DownloadTask> = Vec::new();
-    let mut warnings: Vec<String> = Vec::new();
-    // Normalised local path -> original name, so a collision can be
-    // reported with both sides of the clash.
-    let mut taken_paths: HashMap<String, String> = HashMap::new();
-
-    // The locally saved "<id>_files.xml" occupies the item root too: an
-    // entry that sanitizes to that name (differing only by case on
-    // case-insensitive filesystems) would silently overwrite the metadata.
-    let xml_file_name = xml_file_name_of(base_url);
-    taken_paths.insert(local_path_key(xml_file_name), xml_file_name.to_string());
-
-    for file in files {
-        // Percent-encode the name first so '?' / '#' / '%' characters in
-        // it cannot split the URL into query or fragment components.
-        let encoded_name = encode_download_path(&file.name);
-
-        if encoded_name.is_empty() {
-            warnings.push(format!(
-                "{} {} {} (empty name)",
-                "⚠".yellow().bold(),
-                "Skipped:".yellow(),
-                file.name.dimmed()
-            ));
-            continue;
-        }
-
-        let absolute_url = base_url.join(&encoded_name)?;
-
-        // Sanitize filename for filesystem compatibility
-        let (sanitized_name, was_modified) = sanitize_filename(&file.name);
-
-        // Collect a warning line if the filename was modified
-        if was_modified {
-            warnings.push(format!(
-                "{} {} {} → {}",
-                "⚠".yellow().bold(),
-                "Sanitized:".yellow(),
-                file.name.dimmed(),
-                sanitized_name.bold()
-            ));
-            sanitized_count += 1;
-        }
-
-        // Two different remote names may sanitize to the same local path
-        // (e.g. "file:1.mp4" and "file_1.mp4"); downloading both would
-        // overwrite the earlier file, so keep the first entry and skip the
-        // rest.
-        // The key normalises case on case-insensitive filesystems, so
-        // "Report.PDF" and "report.pdf" collide there but not on Linux.
-        let path_key = local_path_key(&sanitized_name);
-        if let Some(first_name) = taken_paths.get(&path_key) {
-            warnings.push(format!(
-                "{} {} {} collides with {} at {} — the later entry is skipped",
-                "⚠".yellow().bold(),
-                "Collision:".yellow(),
-                file.name.dimmed(),
-                first_name.dimmed(),
-                sanitized_name.bold()
-            ));
-            continue;
-        }
-        taken_paths.insert(path_key, file.name.clone());
-
-        tasks.push(DownloadTask {
-            url: absolute_url.to_string(),
-            file_path: sanitized_name,
-            expected_md5: file.md5,
-            expected_size: file.size,
-            expected_mtime: file.mtime,
-        });
-    }
-
-    Ok(DownloadPlan {
-        tasks,
-        sanitized_count,
-        warnings,
-    })
 }
 
 /// Return formatted file rows for `--list` output.
@@ -448,6 +317,7 @@ async fn run(cli: &Cli) -> Result<()> {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use ia_get::archive_metadata::XmlFile;
     use ia_get::utils::validate_archive_url;
 
     #[test]
@@ -512,142 +382,5 @@ mod tests {
             list_summary(&files),
             "3 files, 3.00MB total known size, 1 unknown size"
         );
-    }
-
-    #[test]
-    fn files_to_download_excludes_xml_self_reference() {
-        let files = vec![
-            xml_file("item1_files.xml", Some(123)),
-            xml_file("scan.jpg", Some(456)),
-        ];
-
-        let result = files_to_download(files, "item1_files.xml");
-
-        assert_eq!(result.len(), 1);
-        assert_eq!(result[0].name, "scan.jpg");
-    }
-
-    #[test]
-    fn files_to_download_keeps_all_when_no_self_reference() {
-        let files = vec![xml_file("scan.jpg", Some(456)), xml_file("notes.txt", None)];
-
-        let result = files_to_download(files, "item1_files.xml");
-
-        assert_eq!(result.len(), 2);
-    }
-
-    #[test]
-    fn plan_download_tasks_skips_entries_with_empty_names() {
-        let base = Url::parse("https://archive.org/download/item1/item1_files.xml").unwrap();
-        let files = vec![
-            xml_file("", Some(10)),
-            xml_file("//", None),
-            xml_file("ok.bin", Some(5)),
-        ];
-        let plan = plan_download_tasks(files, &base).expect("plan must build");
-
-        assert_eq!(plan.sanitized_count, 0);
-        assert_eq!(
-            plan.tasks.len(),
-            1,
-            "empty-name entries must be skipped, not joined to the base URL"
-        );
-        assert_eq!(plan.tasks[0].file_path, "ok.bin");
-        assert_eq!(
-            plan.tasks[0].url,
-            "https://archive.org/download/item1/ok.bin"
-        );
-        assert_eq!(
-            plan.warnings.len(),
-            2,
-            "each skipped entry must leave a warning line"
-        );
-    }
-
-    #[test]
-    fn plan_download_tasks_skips_sanitized_name_collisions() {
-        let base = Url::parse("https://archive.org/download/item1/item1_files.xml").unwrap();
-        // Both names sanitize to "file_1.mp4"; the later entry must not
-        // overwrite the earlier file.
-        let files = vec![
-            xml_file("file:1.mp4", Some(1)),
-            xml_file("file_1.mp4", Some(2)),
-        ];
-        let plan = plan_download_tasks(files, &base).expect("plan must build");
-
-        assert_eq!(plan.sanitized_count, 1, "only the colon name is sanitized");
-        assert_eq!(plan.tasks.len(), 1, "the colliding entry must be skipped");
-        assert_eq!(
-            plan.tasks[0].url,
-            "https://archive.org/download/item1/file%3A1.mp4"
-        );
-        assert_eq!(plan.tasks[0].file_path, "file_1.mp4");
-    }
-
-    #[test]
-    fn plan_download_tasks_case_collision_follows_filesystem() {
-        // Names differing only by case: on case-insensitive filesystems
-        // (Windows, default macOS) the later entry must be skipped, on
-        // case-sensitive ones (Linux) both names are distinct paths.
-        let base = Url::parse("https://archive.org/download/item1/item1_files.xml").unwrap();
-        let files = vec![
-            xml_file("Report.PDF", Some(1)),
-            xml_file("report.pdf", Some(2)),
-        ];
-        let plan = plan_download_tasks(files, &base).expect("plan must build");
-
-        #[cfg(target_os = "linux")]
-        {
-            assert_eq!(
-                plan.tasks.len(),
-                2,
-                "differing case must stay two files on Linux"
-            );
-        }
-        #[cfg(not(target_os = "linux"))]
-        {
-            assert_eq!(
-                plan.tasks.len(),
-                1,
-                "the case-colliding entry must be skipped"
-            );
-            assert_eq!(plan.tasks[0].file_path, "Report.PDF");
-        }
-    }
-
-    #[test]
-    fn plan_download_tasks_protects_saved_xml_name() {
-        // "<id>_files.xml" is saved locally as file #1: an entry that
-        // sanitizes to the same name must not overwrite the metadata.
-        let base = Url::parse("https://archive.org/download/item1/item1_files.xml").unwrap();
-        let files = vec![
-            xml_file("Item1_Files.XML", Some(1)),
-            xml_file("scan.jpg", Some(2)),
-        ];
-        let plan = plan_download_tasks(files, &base).expect("plan must build");
-
-        #[cfg(target_os = "linux")]
-        assert_eq!(
-            plan.tasks.len(),
-            2,
-            "differing case stays distinct on case-sensitive filesystems"
-        );
-        #[cfg(not(target_os = "linux"))]
-        {
-            assert_eq!(
-                plan.tasks.len(),
-                1,
-                "the xml-name-colliding entry must be skipped"
-            );
-            assert_eq!(plan.tasks[0].file_path, "scan.jpg");
-        }
-    }
-
-    #[test]
-    fn local_path_key_normalises_case_per_platform() {
-        #[cfg(target_os = "linux")]
-        assert_eq!(local_path_key("A/b.pdf"), "A/b.pdf");
-        #[cfg(not(target_os = "linux"))]
-        assert_eq!(local_path_key("A/b.PDF"), "a/b.pdf");
     }
 }
